Inclusão Controller de Cartao e Ajustar do ID long

See merge request itau-prereq-1/api-cartoes!19

+++ src/main/java/com/br/api/cartao/repositories/CartaoRepository.java
@@ -3,6 +3,6 @@ package com.br.api.cartao.repositories;
 import com.br.api.cartao.models.Cartao;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CartaoRepository extends CrudRepository<Cartao, Integer> {
+public interface CartaoRepository extends CrudRepository<Cartao, Long> {
 
 }
